RA-20 Creating a ViewModel for recipes and an entity for state
