+++ app/src/main/java/com/example/recipesapp/data/recipes.kt
@@ -1,4 +1,8 @@
-package com.example.recipesapp
+package com.example.recipesapp.data
+
+import com.example.recipesapp.model.Category
+import com.example.recipesapp.model.Ingredient
+import com.example.recipesapp.model.Recipe
 
 object STUB {
     private val categories: List<Category> = listOf(
