+++ app/src/main/java/com/example/recipesapp/model/Category.kt
@@ -1,13 +1,12 @@
-package com.example.recipesapp
+package com.example.recipesapp.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class Recipe(
+data class Category(
     val id: Int,
     val title: String,
-    val ingredients: List<Ingredient>,
-    val method: List<String>,
+    val description: String,
     val imageUrl: String,
 ) : Parcelable
