+++ app/src/main/java/com/example/recipesapp/model/Ingredient.kt
@@ -1,4 +1,4 @@
-package com.example.recipesapp
+package com.example.recipesapp.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
