+++ app/src/main/java/com/example/recipesapp/model/Recipe.kt
@@ -1,12 +1,13 @@
-package com.example.recipesapp
+package com.example.recipesapp.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class Category(
+data class Recipe(
     val id: Int,
     val title: String,
-    val description: String,
+    val ingredients: List<Ingredient>,
+    val method: List<String>,
     val imageUrl: String,
 ) : Parcelable
