+++ app/src/main/java/com/example/recipesapp/ui/IngredientsAdapter.kt
@@ -1,9 +1,11 @@
-package com.example.recipesapp
+package com.example.recipesapp.ui
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.example.recipesapp.R
 import com.example.recipesapp.databinding.ItemIngredientBinding
+import com.example.recipesapp.model.Ingredient
 import java.math.BigDecimal
 import java.math.RoundingMode
