+++ app/src/main/java/com/example/recipesapp/ui/MethodAdapter.kt
@@ -1,8 +1,9 @@
-package com.example.recipesapp
+package com.example.recipesapp.ui
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.example.recipesapp.R
 import com.example.recipesapp.databinding.ItemMethodBinding
 
 class MethodAdapter(
