+++ app/src/main/res/layout/fragment_recipe.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/backgroundColor"
     android:orientation="vertical"
-    tools:context=".RecipeFragment">
+    tools:context=".ui.RecipeFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
